+++ frontend/src/App.tsx
@@ -9,6 +9,7 @@ import {
 } from "@mui/joy";
 import {ThemeSwitcher} from "./components/ThemeSwitcher";
 import {HomeButton} from "./components/HomeButton";
+import { ServerList } from "./components/ServerList";
 import {useEffect, useState} from "react";
 import {Server} from "./pages/Server";
 import {IconArrowLeft, IconExternalLink, IconPlus, IconRefresh} from "@tabler/icons-react";
@@ -78,66 +79,12 @@ function App() {
         }).catch((r) => console.error(r))
     }
 
-    function getServerList(serverListType: ServerListType) {
-        return (
-            <>
-                <List className={'flex flex-row flex-wrap items-start'}>
-                {
-                    (servers === null) ? (
-                        <ListItem>
-                            No servers found or failed to find servers
-                        </ListItem>
-                    ) : (
-                        Object.keys(servers).map((key) => {
-                            const index = parseInt(key, 10); // The second argument is the base (radix), 10 for base 10 (decimal)
-
-                            if (isNaN(index)) {
-                                LogError("Parsing server key failed")
-                            }
-
-                            const server = servers[index]; // Parse the key to a number
-                            if(serverListType === ServerListType.LIST){
-                                return (
-                                    <ListItem key={index}>
-                                        <ListItemButton onClick={() => {setActiveServer(index); setDrawerOpen(false)}}>
-                                            {index}: {server.serverAlias? server.serverAlias : "Unnamed Server"}
-                                        </ListItemButton>
-                                    </ListItem>
-                                );
-                            }
-                            else if(serverListType === ServerListType.CARD){
-                                console.log(server)
-                                return (
-                                    <ListItem className={'w-[calc(100%_*_(1/4))] p-[10px]'} key={index}>
-                                        <Card className={'cursor-pointer w-full'} onClick={() => {setActiveServer(index)}}>
-                                            <div className={'px-2'}>
-                                                <div className={'text-xl font-bold break-all'}>{server.serverAlias? server.serverAlias : "Unnamed Server"}</div>
-                                                <div className={'text-lg mb-2'}>{server.ipAddress}:{server.queryPort}</div>
-
-                                                <div>Players: 0/40</div> {/* this still needs to be implemented! */}
-                                                <div>Status: Running</div> {/* this still needs to be implemented! */}
-                                            </div>
-                                        </Card>
-                                    </ListItem>
-                                );
-                            }
-                        })
-                    )
-                }
-                </List>
-                {serverListType === ServerListType.CARD ? <Button className={'ml-[10px]'} onClick={() => handleCreateNewServerClicked()}><IconPlus/> Create new server</Button> : <></>}
-            </>
-        )
-    }
-
     const ServerDrawer = (
             <Drawer open={drawerOpen} onClose={() => setDrawerOpen(false)} size="md">
                 <ModalClose/>
 
                 <DialogTitle>Servers:</DialogTitle>
-                <List>
-                    {getServerList(ServerListType.LIST)}
-                </List>
+                <ServerList serverListType={ServerListType.LIST} servers={servers} setActiveServer={setActiveServer} setDrawerOpen={setDrawerOpen} handleCreateNewServerClicked={handleCreateNewServerClicked} />
                 <Divider></Divider>
                 <DialogActions>
                     <List>
@@ -164,7 +111,7 @@ function App() {
             mainUi = (
                 <div className={'row-span-5 m-5'}>
                     <h2>Select a server:</h2>
-                    {getServerList(ServerListType.CARD)}
+                    <ServerList serverListType={ServerListType.CARD} servers={servers} setActiveServer={setActiveServer} setDrawerOpen={setDrawerOpen} handleCreateNewServerClicked={handleCreateNewServerClicked} />
                 </div>
             )
         }
